Broke the program into different files
